business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/Reset.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class Reset implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+* ");
        messageString.append(command.getParameter("Reset"));
        messageString.append("\n");
        return messageString;
    }
}
